feat: Set up web app shell
This sets up the web app structure and creates an app shell with a common app
header. The web app uses Firebase Auth to authenticate users when they navigate
to /auth and click the "Sign In with Google" button. There is a /auth/signOut
route that signs out the user and redirects them to the / route. Firebase
initialization takes place in pages/_app.tsx and lib/firebase-client.ts. The
user profile page at /profile displays information for the currently signed in
user. lib/types.d.ts is meant to be the global types file for the project, but
this may change if there are enough data types to warrant refactoring.
Authentication functions are grouped in an AuthContext. A ProfileDialog
(triggered by an AppHeader) contains a useful dialog for accessing user-specific
routes and signing in and out of the app.