feat: Set up PWA service worker and manifest
The public/manifest.json file contains information used during app install. The
theme color for the app is #5D5FEF. At some point, this needs to be reflected in
the theming system. The next-pwa library automatially sets up the service worker
that responds to offline requests. The _app.tsx file now contains HTML meta
information in the head of all pages. The service worker is disabled by default
in development builds. To enable it, create a ENABLE_PWA environment variable
and set it to true.