feat: Connect event registration form to firestore & emails confirmation
This adds controlled components and input validation to the hackathon
registration form. When the form is submitted, a check is run to see if the user
is signed in, then the data is sent to the '/api/applications' route, which
stores the data in firestore. After this, the data is sent to a newly created
route called '/api/email', which sends an email confirmation with the form
responses to the user. Additionally, the 'dietary' field type in
/lib/types.d.ts was changed from 'string' to 'Array<string>' so multiple dietary
restrictions can be reported and stored in firestore. TODO: Upload resume to
Firebase Cloud Storage