feat(registration): Implemented QR generation and reading prelims
Made QR Codes and QR Code Readers individual autonomous components.
Perliminaries for QR generation and QR reading are complete; the only things
left are to integrate reading into the UI and hook up the required callback to
the user data, allowing check-in, food & merch claims, and event registration to
be fully compatible with HackerTags.

feat #6