fix(calendar): Only render track groups if there are tracks available
Since actualy track instances (like "General", etc) are computed from the
available appointments, the scheduler can't form groups if there are no
appointments available because it never expects there to be no instances of a
group when a group is specified. In reality, this is a bug in the scheduler
component itself, and should be fixed a different way, perhaps by patching React
Scheduler itself, but this will fix it for now by not attempting to render
groupings when there are no groupings available.

fixes #191