fix(calendar): Fix nested `p` tag error
Only "phrasing content" is allowed within a `<p>` tag. The event
description is being rendered as several paragraphs, so the parent
element cannot be a `<p>` tag as that would cause `<p>` tags to be
nested within the parent `<p>` tag. This was causing an annoying console
error about nested `<p>` tags. Using a semantic grouping element
(`<article>`, in this case) is the logical fix.